dispenser: Provide initial template
Add a Dispenser class with TODOs for how to finish the example.
Also add a simple autonomous and teleop.
In autonomous, it intakes and shoots at full speed.
In teleop, use an X-box controller to control it with the
following layout:
X - intake
Y - shoot
B - spin shooter motors (but keep intake/indexer stopped)
Defaults to all motors off when nothing is pressed.